+++ ft_memcpy.c
@@ -6,16 +6,16 @@
 /*   By: owatanab <owatanab@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/21 18:14:16 by owatanab          #+#    #+#             */
-/*   Updated: 2023/06/22 13:52:52 by owatanab         ###   ########.fr       */
+/*   Updated: 2023/06/25 13:25:08 by owatanab         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memcpy(void *restrict dst, const void *restrict src, unsigned n)
+void	*ft_memcpy(void *restrict dst, const void *restrict src, size_t n)
 {
-	unsigned	i;
-	char		*s;
+	size_t	i;
+	char	*s;
 
 	char *restrict t;
 	i = -1;
