+++ ft_strnstr.c
@@ -6,16 +6,16 @@
 /*   By: owatanab <owatanab@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/22 11:34:45 by owatanab          #+#    #+#             */
-/*   Updated: 2023/06/22 14:44:16 by owatanab         ###   ########.fr       */
+/*   Updated: 2023/06/25 14:15:15 by owatanab         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_strnstr(const char *haystack, const char *needle, unsigned len)
+char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	unsigned	i;
-	int			j;
+	size_t	i;
+	int		j;
 
 	j = 0;
 	i = 0;
